requirements.txt
Django
numpy
pyqrcode
django-phonenumber-field
djangorestframework
django-registration
requests
pytest
